refactor html.js global IIEF
