+++ src/utils/IIEF/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export { default as themeOper } from './themeOper';
